+++ lib/js_of_ocaml/js.mli
@@ -685,6 +685,19 @@ val string_of_error : error t -> string
 
 val raise_js_error : error t -> 'a
 
+val exn_with_js_backtrace : exn -> force:bool -> exn
+(** Attach a JavasScript error to an OCaml exception.  if [force = false] and a
+    JavasScript error is already attached, it will do nothing. This function is useful to
+    store and retrieve information about JavaScript stack traces.
+
+    Attaching JavasScript errors will happen automatically when compiling with
+    [--enable with-js-error].
+*)
+
+val js_error_of_exn : exn -> error t opt
+(** Extract a JavaScript error attached to an OCaml exception, if any.  This is useful to
+    inspect an eventual stack strace, especially when sourcemap is enabled. *)
+
 exception Error of error t
 (** The [Error] exception wrap javascript exceptions when caught by OCaml code.
       In case the javascript exception is not an instance of javascript [Error],