+++ lib/js_of_ocaml/js_of_ocaml_stubs.c
@@ -32,6 +32,10 @@ void caml_create_file () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_create_file!\n");
   exit(1);
 }
+void caml_exn_with_js_backtrace () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_exn_with_js_backtrace!\n");
+  exit(1);
+}
 void caml_int64_create_lo_mi_hi () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_int64_create_lo_mi_hi!\n");
   exit(1);
@@ -46,6 +50,10 @@ void caml_js_delete () {
 }
 void caml_js_equals () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_js_equals!\n");
+  exit(1);
+}
+void caml_js_error_of_exception () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_error_of_exception!\n");
   exit(1);
 }
 void caml_js_eval_string () {