+++ CHANGES.md
@@ -1,3 +1,8 @@
+# ??? (??) - ??
+## Features/Changes
+* Compiler: complete support for OCaml 4.12
+* Lib: expose API to attached and retrieve js errors to/from ocaml exceptions
+
 # 3.8.0 (2020-10-21) - London
 ## Features/Changes
 * compiler, ppx_js, ppx_deriving_json: port PPX's to ppxlib (#1041)